% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{AB_CIVARTest}
\alias{AB_CIVARTest}
\title{Legacy alias for \code{AB_CIVARTest()}}
\usage{
AB_CIVARTest(...)
}
\description{
This is kept for backwards compatibility. Please use \code{ab_civar_test()} instead.
}
