% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{ab_mrcivar_test}
\alias{ab_mrcivar_test}
\title{Likelihood ratio test of restrictions in CIVAR models}
\usage{
ab_mrcivar_test(
  R0,
  R1,
  G,
  H,
  h,
  alphaR1,
  alphaR2,
  betaR,
  alpha1,
  alpha2,
  beta1,
  beta2,
  OmegaR,
  Omega1,
  Omega2,
  IC = 1,
  T1,
  St,
  NSt
)
}
\arguments{
\item{R0}{controlled I(0) data matrix}

\item{R1}{controlled I(1) data matrix}

\item{G}{A matrix specifying restrictions on alpha}

\item{H}{A matrix specifying restrictions on beta}

\item{h}{A vector specifying restrictions on alpha}

\item{alphaR1}{Initial value for constrained estimation}

\item{alphaR2}{Initial value for constrained estimation}

\item{betaR}{Initial value for constrained estimation}

\item{alpha1}{Initial value for unconstrained estimation}

\item{alpha2}{Initial value for unconstrained estimation}

\item{beta1}{Initial value for unconstrained estimation}

\item{beta2}{Initial value for unconstrained estimation}

\item{OmegaR}{Initial value for constrained estimation}

\item{Omega1}{Initial value for unconstrained estimation}

\item{Omega2}{Initial value for unconstrained estimation}

\item{IC}{Indicator of BB dimension}

\item{T1}{Number of  used observations}

\item{St}{Regime 1 indicator series}

\item{NSt}{Regime 2 indicator series}
}
\value{
A list containing the constrained estimates and likelihood test results
}
\description{
This function runs Doornik and Boswijk Iteration to estimate restricted alpha and beta in a CIVAR model.
}
