% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{ABC_MRCIVARestm}
\alias{ABC_MRCIVARestm}
\title{Legacy alias for \code{ABC_MRCIVARestm()}}
\usage{
ABC_MRCIVARestm(...)
}
\description{
This is kept for backwards compatibility. Please use \code{abc_mrciva_restm()} instead.
}
