% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{abLRtest2}
\alias{abLRtest2}
\title{Legacy alias for \code{abLRtest2()}}
\usage{
abLRtest2(...)
}
\description{
This is kept for backwards compatibility. Please use \code{ab_l_rtest2()} instead.
}
