% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{accir_fconf_r}
\alias{accir_fconf_r}
\title{Calculate Accumulated Impulse Response Functions}
\usage{
accir_fconf_r(IRF)
}
\arguments{
\item{IRF}{An impulse response function}
}
\value{
The cumulative impulse response function
}
\description{
This function calculates the cumulative impulse response functions from a given set of impulse response functions.
It sums the impulse responses over time to provide a cumulative effect of the shocks.
}
\examples{
# Example usage of accir_fconf_r
# Assuming IRF is a predefined impulse response function arrays
p = matrix(c(2,1,0,0),2,2)
res_d = mrvar_data(n=2,p=p,T=300,S=2,SESVI=1)
max(abs(res_d$Y))
res_e  = mrvar_estimate(res_d)
res_e$Summary

IRF <- irf_mrvar_nm(res_e,nstep=10,comb=NA,irf="gen")
cumulative_response <- accir_fconf_r(IRF)
}
\keyword{internal}
