% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{ACCIRFconfR}
\alias{ACCIRFconfR}
\title{Legacy alias for \code{ACCIRFconfR()}}
\usage{
ACCIRFconfR(...)
}
\description{
This is kept for backwards compatibility. Please use \code{accir_fconf_r()} instead.
}
