% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bo_ao2_t_bo_ao}
\alias{bo_ao2_t_bo_ao}
\title{This function calculates a full rank linear transformation of initial Bo and Ao matrices with independent common trends and idiosyncratic trends.}
\usage{
bo_ao2_t_bo_ao(Bo, Ao, W, S)
}
\arguments{
\item{Bo}{The initial Bo matrix}

\item{Ao}{The initial Ao matrix}

\item{W}{A weighting matrix of the GVAR model}

\item{S}{Number of regimes}
}
\value{
Transformed Bo and Ao matrices
}
\description{
This function calculates a full rank linear transformation of initial Bo and Ao matrices with independent common trends and idiosyncratic trends.
}
\keyword{internal}
