% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bo_ao_ws2_gs}
\alias{bo_ao_ws2_gs}
\title{Assemble regime-specific GVAR coefficient array given a state vector}
\usage{
bo_ao_ws2_gs(Bo, Ao, W, m, n, p, state)
}
\arguments{
\item{Bo}{Parameter arrays of the domestic variables (m x m x p x n x S).}

\item{Ao}{Parameter arrays of the foreign variables (m x m x p x n x S).}

\item{W}{The weighting matrix of the GVAR model (n x n).}

\item{m}{Number of variables per country.}

\item{n}{Number of countries in the GVAR model.}

\item{p}{Lag order.}

\item{state}{Integer vector of length \code{n}, where \code{state[i]} indicates the regime
to use for country \code{i} (values in 1..S).}
}
\value{
An (mn, mn, p) array of autoregressive parameters for the regime-specific GVAR.
}
\description{
Selects, for each country, the regime-specific domestic (\code{Bo}) and foreign (\code{Ao})
coefficient matrices according to \code{state}, then assembles the corresponding stacked
GVAR coefficient array using the weighting matrix \code{W}. This is a convenience wrapper
around \code{bo_ao_w2_g()} for multi-regime GVAR parameters.
}
\examples{
set.seed(1)
n <- 3
m <- 2
p <- 2
S <- 2

Bo <- array(rnorm(m * m * p * n * S), dim = c(m, m, p, n, S))
Ao <- array(rnorm(m * m * p * n * S), dim = c(m, m, p, n, S))
W  <- matrix(1, n, n)
diag(W) <- 0
W <- W / rowSums(W)

state <- c(1, 2, 1)  # choose regime per country
Gs <- bo_ao_ws2_gs(Bo = Bo, Ao = Ao, W = W, m = m, n = n, p = p, state = state)
dim(Gs)

}
\keyword{internal}
