% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{BoAo2TBoAo}
\alias{BoAo2TBoAo}
\title{Legacy alias for \code{BoAo2TBoAo()}}
\usage{
BoAo2TBoAo(...)
}
\description{
This is kept for backwards compatibility. Please use \code{bo_ao2_t_bo_ao()} instead.
}
