% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{BoAoWs2Gs}
\alias{BoAoWs2Gs}
\title{Legacy alias for \code{BoAoWs2Gs()}}
\usage{
BoAoWs2Gs(...)
}
\description{
This is kept for backwards compatibility. Please use \code{bo_ao_ws2_gs()} instead.
}
