% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{CCIVARData}
\alias{CCIVARData}
\title{Legacy alias for \code{CCIVARData()}}
\usage{
CCIVARData(...)
}
\description{
This is kept for backwards compatibility. Please use \code{ccivar_data()} instead.
}
