% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{CCIVARest}
\alias{CCIVARest}
\title{Legacy alias for \code{CCIVARest()}}
\usage{
CCIVARest(...)
}
\description{
This is kept for backwards compatibility. Please use \code{cciva_rest()} instead.
}
