% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cia2_a}
\alias{cia2_a}
\title{Convert VECM coefficient arrays to VAR coefficient arrays in levels (foreign block)}
\usage{
cia2_a(CIA)
}
\arguments{
\item{CIA}{A coefficient array of a VECM (m x m x L).}
}
\value{
An (m, m, L) array containing the coefficient matrices of the corresponding
VAR in levels.
}
\description{
Transforms an array of VECM coefficient matrices into the corresponding VAR
coefficient matrices in levels. Both input and output are 3D arrays, where the
third dimension indexes lags.
}
\examples{
# Example with m = 2 variables and L = 3 lags in the array
set.seed(1)
m <- 2
L <- 3
CIA <- array(rnorm(m * m * L), dim = c(m, m, L))
A <- cia2_a(CIA)
dim(A)
A[,,1]

}
\keyword{internal}
