% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cib3_b}
\alias{cib3_b}
\title{Convert VECM coefficient arrays to VAR coefficient arrays}
\usage{
cib3_b(CIB)
}
\arguments{
\item{CIB}{Three-dimensional array of VECM coefficients.}
}
\value{
A three-dimensional array of VAR coefficients.
}
\description{
Transform an array of VECM coefficients into the corresponding VAR
coefficients in levels.
}
\examples{
# Example with m = 3 variables and L = 2 (so VECM lag is L-1 = 1)
set.seed(1)
m <- 3
L <- 2
CIB <- array(rnorm(m * m * L), dim = c(m, m, L))
B <- cib3_b(CIB)
dim(B)
B[,,1]

}
\keyword{internal}
