% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{CIGVAR_Select}
\alias{CIGVAR_Select}
\title{Legacy alias for \code{CIGVAR_Select()}}
\usage{
CIGVAR_Select(...)
}
\description{
This is kept for backwards compatibility. Please use \code{cigvar_select_()} instead.
}
