% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{CIGVARData}
\alias{CIGVARData}
\title{Legacy alias for \code{CIGVARData()}}
\usage{
CIGVARData(...)
}
\description{
This is kept for backwards compatibility. Please use \code{cigvar_data()} instead.
}
