% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{CIGVARest}
\alias{CIGVARest}
\title{Legacy alias for \code{CIGVARest()}}
\usage{
CIGVARest(...)
}
\description{
This is kept for backwards compatibility. Please use \code{cigvar_estimate()} instead.
}
