% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{CIVARData}
\alias{CIVARData}
\title{Legacy alias for \code{CIVARData()}}
\usage{
CIVARData(...)
}
\description{
This is kept for backwards compatibility. Please use \code{civar_data()} instead.
}
