% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{CIVARest}
\alias{CIVARest}
\title{Legacy alias for \code{CIVARest()}}
\usage{
CIVARest(...)
}
\description{
This is kept for backwards compatibility. Please use \code{civar_estimate()} instead.
}
