% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{CIVARTest}
\alias{CIVARTest}
\title{Legacy alias for \code{CIVARTest()}}
\usage{
CIVARTest(...)
}
\description{
This is kept for backwards compatibility. Please use \code{civar_test()} instead.
}
