% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{CVECM2CVAR}
\alias{CVECM2CVAR}
\title{Legacy alias for \code{CVECM2CVAR()}}
\usage{
CVECM2CVAR(...)
}
\description{
This is kept for backwards compatibility. Please use \code{cvecm2_cvar()} instead.
}
