% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{girf_MRCIGVAR_RM}
\alias{girf_MRCIGVAR_RM}
\title{Legacy alias for \code{girf_MRCIGVAR_RM()}}
\usage{
girf_MRCIGVAR_RM(...)
}
\description{
This is kept for backwards compatibility. Please use \code{girf_mrcigvar_rm()} instead.
}
