% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{girf_mrcigvar_rm_cb}
\alias{girf_mrcigvar_rm_cb}
\title{Compute MRCIGVAR generalized impulse responses with confidence bands}
\usage{
girf_mrcigvar_rm_cb(
  res,
  shock,
  R,
  nstep,
  Omega_hist = NA,
  resid_method = "parametric",
  conf,
  N
)
}
\arguments{
\item{res}{A fitted MRCIGVAR object, typically returned by
\code{mrcigvar_estimate()}.}

\item{shock}{Numeric vector of length \code{m * n} defining the impulse.}

\item{R}{Integer scalar giving the number of simulation runs used to average
over random effects.}

\item{nstep}{Integer scalar giving the number of response horizons.}

\item{Omega_hist}{Matrix of initial values used to start the simulation.}

\item{resid_method}{Character string specifying how residual draws are
generated. Supported values are \code{"resid"} and \code{"parametric"}.}

\item{conf}{Numeric vector of tail probabilities used to form the bootstrap
confidence interval.}

\item{N}{Integer scalar giving the number of bootstrap replications.}
}
\value{
A four-dimensional array of dimension \verb{(m * n, m * n, nstep + 1, length(conf) + 1)} containing the point estimate and confidence bounds.
}
\description{
Compute generalized impulse response functions (GIRFs) for an estimated
MRCIGVAR model and bootstrap confidence bands for a user-specified shock.
}
\details{
For a shock vector \code{SHCK}, the GIRF is defined as
\code{mean(Y(SHCK)) - mean(Y(resid))}.

See H.H. Pesaran and Y. Shin (1998) Generalized impulse response analysis in linear multivariate
models, Economics Letters, 58(1) p. 17-29.
}
\examples{
m = 2
n = 3
p = c(2,2,2,2,2,2,2,2,2,2,0,0,0,0,0,2,2,2,2,2,2,2,2,2,2,0,0,0,0,0)
dim(p) = c(5,3,2)
p = p[1:n,,]
p[,1,] = 3
p[,2,] = 2

TH = c(1:n)*0
dim(TH) = c(1,n)
SESVI = rep(1,3,5)
r = rep(1,n)

## case of n = 3, m = 2, S = 2    m: number of variables, n: number of countries
res_d <- mrcigvar_data(m=2, n=3, p=p, TH=TH, T=200, S=2, SESVI=c(1,3,5), r=rep(1,3), Ncommtrend=1)
max(abs(res_d$Y))
spectral_radius(res_d$Go[,,,2])
spectral_radius(res_d$Go[,,,1])
res_e <- mrcigvar_estimate(res=res_d)

spectral_radius(res_e$Go[,,,2])
spectral_radius(res_e$Go[,,,1])

plot(ts(res_d$Y))

if (!((max(Mod(spectral_radius(res_e$Go[,,,1])))>1)|
 (max(Mod(spectral_radius(res_e$Go[,,,2])))>1)) ) {
  GIRF <- girf_mrcigvar_rm(res=res_e, shock=c(1,1,1,1,1,1), R=100, nstep=10, Omega_hist=NA,
  resid_method="parametric")
  GIRF_CB <- girf_mrcigvar_rm_cb(res=res_e, shock=c(1,1,1,1,1,1), R=100, nstep=10, Omega_hist=NA,
  resid_method="parametric", conf=c(0.05,0.95), N=100)
  IRF_g = plot_irf(GIRF_CB)
}

}
