% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{girf_MRCIVAR_RM}
\alias{girf_MRCIVAR_RM}
\title{Legacy alias for \code{girf_MRCIVAR_RM()}}
\usage{
girf_MRCIVAR_RM(...)
}
\description{
This is kept for backwards compatibility. Please use \code{girf_mrcivar_rm()} instead.
}
