% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{girf_MRVAR_RM}
\alias{girf_MRVAR_RM}
\title{Legacy alias for \code{girf_MRVAR_RM()}}
\usage{
girf_MRVAR_RM(...)
}
\description{
This is kept for backwards compatibility. Please use \code{girf_mrvar_rm()} instead.
}
