% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{gvar_data}
\alias{gvar_data}
\title{Generate GVAR synthetic data}
\usage{
gvar_data(
  m,
  n,
  p,
  T,
  W = NA,
  r_npo = NA,
  Ao = NA,
  Bo = NA,
  Co = NA,
  Uo = NA,
  Sigmao = NA,
  type = NA,
  X = NA,
  mu = NA
)
}
\description{
Creates a GVAR dataset under the package’s model specification.
}
\examples{
\dontrun{
p = (1:12)*0; dim(p) = c(4,3);p[,1] = 2; p[,2]=1;   p[,3]=1; p[2,2]=2;
p    ## country-wise lag specification

res_d = gvar_data(m=2,n=4,p=p,T=200,type="exog0",X=X)
res_e = gvar_estimate(res = res_d)
res_e$Summary

IRF_CB = irf_gvar_cb(res_e,nstep=10,comb=NA,irf="gen",runs=200,conf=c(0.05,0.95))
dim(IRF_CB)
IRF_g = plot_irf(IRF_CB,Names=NA,response=c(1,4),impulse=c(1,2,3,4), ncol=4)
}

}
