% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{gvar_estimate}
\alias{gvar_estimate}
\title{p = (1:12)*0; dim(p) = c(4,3);p\link{,1} = 2; p\link{,2}=1;   p\link{,3}=1; p\link{2,2}=2;
p}
\usage{
gvar_estimate(res)
}
\description{
res_d = gvar_data(m=2,n=4,p=p,T=200,type="exog0",X=X)
res_e = gvar_estimate(res = res_d)
res_e$Summary
}
\details{
IRF_CB = irf_gvar_cb(res_e,nstep=10,comb=NA,irf="gen",runs=200,conf=c(0.05,0.95))
dim(IRF_CB)
IRF_g = plot_irf(IRF_CB,Names=NA,response=c(1,4),impulse=c(1,2,3,4), ncol=4)
}
