% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{GVAR_Select}
\alias{GVAR_Select}
\title{Legacy alias for \code{GVAR_Select()}}
\usage{
GVAR_Select(...)
}
\description{
This is kept for backwards compatibility. Please use \code{gvar_select_()} instead.
}
