% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{gvar_select_}
\alias{gvar_select_}
\title{GVAR Model Selection Function}
\usage{
gvar_select_(res = res, L_V = L_V, I = I)
}
\arguments{
\item{res}{A GVAR object generated from gvar_data}

\item{L_V}{A 2 components vector containing the maxima of the domestic lag and the foreign lag, respectively}

\item{I}{Index of the country equation under investigation}
}
\value{
A matrix with different lag specifications and the corresponding information criteria (AIC and BIC)
}
\description{
This function performs model selection for the GVAR (Global Vector Autoregression) framework by calculating AIC and BIC criteria for various lag specifications of domestic and foreign variables for a specific country equation.
}
\examples{
n = 4
p = (1:12)*0; dim(p) = c(4,3);p[,1] = 2; p[,2]=1; p[2:3,2] = 2
res_d = gvar_data(m=2,n=4,p=p,T=4000,type="const")

I = 3
L_V = c(4,4)
res_d$p
GVARSelect = gvar_select_(res=res_d,L_V=c(4,4),I=2)
GVARSelect[which.min(GVARSelect[,3]),]
GVARSelect[which.min(GVARSelect[,4]),]

}
