% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{irf_B_sigma}
\alias{irf_B_sigma}
\title{Legacy alias for \code{irf_B_sigma()}}
\usage{
irf_B_sigma(...)
}
\description{
This is kept for backwards compatibility. Please use \code{irf_from_params()} instead.
}
