% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{irf_CIGVAR_CB}
\alias{irf_CIGVAR_CB}
\title{Legacy alias for \code{irf_CIGVAR_CB()}}
\usage{
irf_CIGVAR_CB(...)
}
\description{
This is kept for backwards compatibility. Please use \code{irf_cigvar_cb()} instead.
}
