% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{irf_CIVAR_CB}
\alias{irf_CIVAR_CB}
\title{Legacy alias for \code{irf_CIVAR_CB()}}
\usage{
irf_CIVAR_CB(...)
}
\description{
This is kept for backwards compatibility. Please use \code{irf_civar_cb()} instead.
}
