% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf_from_params}
\alias{irf_from_params}
\title{Generate Impulse Response Functions from Model Parameters}
\usage{
irf_from_params(
  B,
  sigma,
  nstep,
  comb,
  irf = c("gen", "chol", "chol1", "gen1", "genN1", "comb1", "smat", "concerts1",
    "PTdecomp", "ABSVAR", "irfX"),
  G = NA,
  A0 = NA,
  B0 = NA,
  smat = NA,
  Xshk = NA
)
}
\arguments{
\item{B}{An (nxnxp) coefficients array of an n-dimensional VAR(p) model}

\item{sigma}{The covariance matrix of the VAR(p) model}

\item{nstep}{Number of steps of the impulse response function}

\item{comb}{An n-vector of weights of the coordinated policy actions}

\item{irf}{Type of impulse response function}

\item{G}{The matrix used in the permanent and transitory decomposition}

\item{A0}{The matrix for A/B identification in VAR}

\item{B0}{The matrix for A/B identification in VAR}

\item{smat}{An explicit decomposition matrix that defines a structural shock.}

\item{Xshk}{The shock matrix of a one unit exogenous shock}
}
\description{
This function generates impulse response functions for various types of models including VAR, CIVAR, MRVAR, MRCIVAR, GVAR, CIGVAR, MRGVAR, and MRCIGVAR.
It also provides functionalities to calculate global, regional, and country-specific shocks, as well as global and regional responses and coordinated policy actions.
}
