% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf_gloabal_response_cb}
\alias{irf_gloabal_response_cb}
\title{This function calculates global or regional responses from a set of bootstrapped impulse response functions and a weighting matrix for the aggregation of the global responses.}
\usage{
irf_gloabal_response_cb(IRF_CB, comb_all)
}
\arguments{
\item{IRF_CB}{an output of irf_mrgvar_cb}

\item{comb_all}{a weighting matrix for the aggregation of the global responses}
}
\value{
a list containing the global impulse response functions and the accumulated global impulse response functions.
}
\description{
This function calculates global or regional responses from a set of bootstrapped impulse response functions and a weighting matrix for the aggregation of the global responses.
}
