% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{irf_GloabalResponse_CB}
\alias{irf_GloabalResponse_CB}
\title{Legacy alias for \code{irf_GloabalResponse_CB()}}
\usage{
irf_GloabalResponse_CB(...)
}
\description{
This is kept for backwards compatibility. Please use \code{irf_gloabal_response_cb()} instead.
}
