% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf_global_response_cb}
\alias{irf_global_response_cb}
\title{Calculate Global Impulse Response Functions with Confidence Bands}
\usage{
irf_global_response_cb(IRF_CB, comb_all)
}
\arguments{
\item{IRF_CB}{an output of irf_mrgvar_cb}

\item{comb_all}{a weighting matrix for the aggregation of the global responses}
}
\value{
a list containing the global impulse response functions and the accumulated global impulse response functions.
}
\description{
This function calculates global or regional responses from a set of bootstrapped impulse response functions
and a weighting matrix for the aggregation of the global responses. It returns both the global impulse response
functions and their accumulated versions, along with confidence intervals.
}
\examples{
# Example usage of irf_global_response_cb
# Assuming IRF_CB is the output from irf_mrgvar_cb and comb_all is a predefined weighting matrix
# Define a simple "comb_all" (weighting matrix) for the global aggregation
# Equal-weight "comb_all" across 2 countries (two identical columns)
\dontrun{
comb_all = matrix(c(0.5, 0.5, 0.5, 0.5), nrow = 2, ncol = 2)
p = matrix(c(2,1,0,0),2,2)
res_d = mrvar_data(n=2,p=p,T=300,S=2,SESVI=1)
max(abs(res_d$Y))
res_e  = mrvar_estimate(res_d)
res_e$Summary

IRF_CB <- irf_mrvar_nm_cb(res_e,nstep=10,comb=NA,irf="gen",runs=200,conf=c(0.05,0.90))
result <- irf_global_response_cb(IRF_CB, comb_all)
}
}
