% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{IRF_graph}
\alias{IRF_graph}
\title{Legacy alias for \code{IRF_graph()}}
\usage{
IRF_graph(...)
}
\description{
This is kept for backwards compatibility. Please use \code{plot_irf()} instead.
}
