% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{irf_GVAR_CB}
\alias{irf_GVAR_CB}
\title{Legacy alias for \code{irf_GVAR_CB()}}
\usage{
irf_GVAR_CB(...)
}
\description{
This is kept for backwards compatibility. Please use \code{irf_gvar_cb()} instead.
}
