% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf_mrcigvar}
\alias{irf_mrcigvar}
\title{Generate Impulse Response Functions for MRCIGVAR Models}
\usage{
irf_mrcigvar(
  res,
  nstep,
  comb,
  state = state,
  irf = c("gen", "chol", "chol1", "gen1", "comb1"),
  sigmaNPDS = NA
)
}
\arguments{
\item{res}{a list of estimated MRCIGVAR as output of mrcigvar_estimate}

\item{nstep}{the length of impulse response function}

\item{comb}{a vector specify the concerted action in policy-simulation impulse response function}

\item{state}{an n vector specifying the specific state for each country}

\item{irf}{types of the impulse response irf=c("gen","chol","chol1","gen1","comb1")}

\item{sigmaNPDS}{the state-dependent covariance matrix}
}
\value{
an (mn,mn,nstep) array containing the IRF with columns representing the impulses and rows the responses
}
\description{
This function computes the impulse response functions (IRFs) of an estimated MRCIGVAR model,
allowing for various types of IRFs and handling different configurations of shocks and responses
across multiple regimes and countries.
}
\examples{
m = 2
n = 3
p = c(2,2,2,2,2,2,2,2,2,2,0,0,0,0,0,2,2,2,2,2,2,2,2,2,2,0,0,0,0,0); dim(p) = c(5,3,2)
p = p[1:n,,]; p[,1,] = 3; p[,2,] = 2

TH = c(1:n)*0; dim(TH) = c(1,n)
SESVI=rep(1,3,5)
r  = rep(1,n)

## case of n = 3, m = 2, S = 2    m: number of variables, n: number of countries

res_d <- mrcigvar_data(m=2,n=3,p=p,TH=TH,T=300,S=2, SESVI=c(1,3,5),r=rep(1,3),Ncommtrend=1)
max(abs(res_d$Y))
spectral_radius(res_d$Go[,,,2])
spectral_radius(res_d$Go[,,,1])
res_e  = mrcigvar_estimate(res=res_d)

res_e$Summary

IRFF2 = irf_mrcigvar(res=res_e,nstep=10,comb=NA,state=c(2,2,2),irf='gen')
plot(IRFF2[1,1,],type="l")

}
\keyword{internal}
