% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{irf_MRCIGVAR_CB}
\alias{irf_MRCIGVAR_CB}
\title{Legacy alias for \code{irf_MRCIGVAR_CB()}}
\usage{
irf_MRCIGVAR_CB(...)
}
\description{
This is kept for backwards compatibility. Please use \code{irf_mrcigvar_cb()} instead.
}
