% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf_mrvar_cb}
\alias{irf_mrvar_cb}
\title{This function calculates the regime specific impulse response functions of an estimated MRVAR(n,p,S).}
\usage{
irf_mrvar_cb(
  res_e,
  nstep,
  comb,
  irf = c("gen", "chol", "chol1", "gen1", "comb1", "irfX"),
  Xshks = NA,
  runs = 200,
  conf = c(0.05, 0.95)
)
}
\arguments{
\item{res_e}{an output of MRVARest.}

\item{nstep}{the length of impulse response function}

\item{comb}{a vector specify the concerted action in policy-simulation impulse response function}

\item{irf}{types of the impulse response functions. irf=c("gen","chol","chol1","gen1","comb1"), gen for generalized IRF with one standard deviation shocks, gen1 for generalized IRF with one unit impulse, chol for IRF with Cholezky decomposition of the covariance matrix, chol1 for Cholezky decomposition with one unit impulse, comb1 for concerted action with one unit impulse in GVAR.}

\item{Xshks}{number of selected exogenous shocks}

\item{runs}{number of simulation runs}

\item{conf}{a two components vector containing the tail probabilities of the bootstrap confidence interval.}
}
\value{
an (n,n,nstep,3,2) array containing the IRF of the two regimes. The IRF columns represent the impulse and the rows the responses.
}
\description{
This function calculates the regime specific impulse response functions of an estimated MRVAR(n,p,S).
}
\examples{
p = matrix(c(2,1,0,0),2,2)
res_d = mrvar_data(n=2,p=p,T=300,S=2,SESVI=1)
max(abs(res_d$Y))
res_e  = mrvar_estimate(res_d)
res_e$Summary


IRF_CB = irf_mrvar_cb(res_e,nstep=20,comb=NA,irf="gen1",runs=100,conf=c(0.05,0.90))
IRF_list1 <-plot_irf(IRF_CB[[1]])
IRF_list2 <-plot_irf(IRF_CB[[2]])

}
