% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf_var_cb}
\alias{irf_var_cb}
\title{Calculate Bootstrapped Confidence Intervals for Impulse Response Functions of VAR Models}
\usage{
irf_var_cb(
  res,
  nstep,
  comb,
  irf = c("gen", "chol", "chol1", "gen1", "comb1", "irfX"),
  runs = 200,
  conf = c(0.05, 0.95),
  smat = NA,
  Xshks = NA
)
}
\arguments{
\item{res}{an object of the output of var_estimate}

\item{nstep}{length of the impulse response functions}

\item{comb}{a vector containing the weights of a combined policy actions}

\item{irf}{types of impulse response functions}

\item{runs}{number of simulation runs in generating the confidence interval of the impulse response functions}

\item{conf}{confidence levels}

\item{smat}{shock to response transformation matrix for the option irf="smat"}

\item{Xshks}{the number of exogenous variables}
}
\value{
an (n, n, nstep, 3) array containing the impulse response functions
}
\description{
This function generates impulse response functions of an estimated VAR(p) model,
providing bootstrapped confidence intervals for the impulse response functions.
}
\examples{
res_d = var_data(n=2,p=2,T=100,type="const")
res_e = var_estimate(res=res_d)
IRF_CB = irf_var_cb(res=res_e,nstep=20, comb=NA, irf = "gen1", runs = 100, conf = c(0.05, 0.95))
IRF_list <- plot_irf(IRF_CB)
res_e$Summary

}
