% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{JointX}
\alias{JointX}
\title{Legacy alias for \code{JointX()}}
\usage{
JointX(...)
}
\description{
This is kept for backwards compatibility. Please use \code{bind_exogenous_by_state()} instead.
}
