% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{MIxCIVARData}
\alias{MIxCIVARData}
\title{Legacy alias for \code{MIxCIVARData()}}
\usage{
MIxCIVARData(...)
}
\description{
This is kept for backwards compatibility. Please use \code{m_ix_civar_data()} instead.
}
