% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{MRCIGVAR_Select}
\alias{MRCIGVAR_Select}
\title{Legacy alias for \code{MRCIGVAR_Select()}}
\usage{
MRCIGVAR_Select(...)
}
\description{
This is kept for backwards compatibility. Please use \code{mrcigvar_select_()} instead.
}
