% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{mrcigvar_select_}
\alias{mrcigvar_select_}
\title{MRCIGVAR Model Selection Function}
\usage{
mrcigvar_select_(res, L_V = L_V, TH_V = TH_V)
}
\arguments{
\item{res}{: an MRCIGVAR object generated from mrcigvar_data or estimated from mrcigvar_estimate.}

\item{L_V}{: a 2 components vector containing the maxima of the domestic lag length and the foreign lag length respectively.}

\item{TH_V}{: a vector containing possible threshold values.}
}
\value{
a matrix with different lag specifications, threshold values and the corresponding information criteria.
}
\description{
This function performs model selection for the MRCIGVAR (Markov Regime-Switching Cointegrated GVAR) framework by calculating AIC, BIC, and other information criteria for various lag specifications and threshold values across regimes and countries.
}
\examples{
m = 2                  ## m: number of variables, n: number of countries
n = 4
p = c(2,2,2,2,2,2,2,2,2,2,0,0,0,0,0,2,2,2,2,2,2,2,2,2,2,0,0,0,0,0); dim(p) = c(5,3,2)
p = p[1:4,,]; p[,1:2,2] = 3

TH = c(1:4)*0; dim(TH) = c(1,4)

res_d <- mrcigvar_data(m=2,n=4,p=p,TH=TH,T=400,S=2, SESVI=c(1,3,5,7),r=rep(1,4))

res_em = mrcigvar_estimate(res=res_d)

L_v = c(3,3)
TH_v = c(0,0.1)
res_d$p

MRCIGVARSelect = mrcigvar_select_(res=res_d,L_V=L_v,TH_V=TH_v)
dim(MRCIGVARSelect)
MRCIGVARSelect[which.min(MRCIGVARSelect[,28]),]

}
