% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{MRCIGVARData}
\alias{MRCIGVARData}
\title{Legacy alias for \code{MRCIGVARData()}}
\usage{
MRCIGVARData(...)
}
\description{
This is kept for backwards compatibility. Please use \code{mrcigvar_data()} instead.
}
