% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{MRCIGVARDataR}
\alias{MRCIGVARDataR}
\title{Legacy alias for \code{MRCIGVARDataR()}}
\usage{
MRCIGVARDataR(...)
}
\description{
This is kept for backwards compatibility. Please use \code{mrcigvar_data_r()} instead.
}
