% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{MRCIGVARest}
\alias{MRCIGVARest}
\title{Legacy alias for \code{MRCIGVARest()}}
\usage{
MRCIGVARest(...)
}
\description{
This is kept for backwards compatibility. Please use \code{mrcigvar_estimate()} instead.
}
