% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{MRCIVAR_Selectm}
\alias{MRCIVAR_Selectm}
\title{Legacy alias for \code{MRCIVAR_Selectm()}}
\usage{
MRCIVAR_Selectm(...)
}
\description{
This is kept for backwards compatibility. Please use \code{mrcivar_select_m()} instead.
}
