% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{MRCIVARDatam}
\alias{MRCIVARDatam}
\title{Legacy alias for \code{MRCIVARDatam()}}
\usage{
MRCIVARDatam(...)
}
\description{
This is kept for backwards compatibility. Please use \code{mrcivar_data_m()} instead.
}
