% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{MRCIVARestm1}
\alias{MRCIVARestm1}
\title{Legacy alias for \code{MRCIVARestm1()}}
\usage{
MRCIVARestm1(...)
}
\description{
This is kept for backwards compatibility. Please use \code{mrcivar_estimatem1()} instead.
}
