% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{MRCVECMest2}
\alias{MRCVECMest2}
\title{Legacy alias for \code{MRCVECMest2()}}
\usage{
MRCVECMest2(...)
}
\description{
This is kept for backwards compatibility. Please use \code{mrcvec_mest2()} instead.
}
