% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{MRCVECMestm}
\alias{MRCVECMestm}
\title{Legacy alias for \code{MRCVECMestm()}}
\usage{
MRCVECMestm(...)
}
\description{
This is kept for backwards compatibility. Please use \code{mrcvec_mestm()} instead.
}
