% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{MRGVAR_Select}
\alias{MRGVAR_Select}
\title{Legacy alias for \code{MRGVAR_Select()}}
\usage{
MRGVAR_Select(...)
}
\description{
This is kept for backwards compatibility. Please use \code{mrgvar_select_()} instead.
}
