% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{MRGVARData}
\alias{MRGVARData}
\title{Legacy alias for \code{MRGVARData()}}
\usage{
MRGVARData(...)
}
\description{
This is kept for backwards compatibility. Please use \code{mrgvar_data()} instead.
}
