% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{MRGVARDataR}
\alias{MRGVARDataR}
\title{Legacy alias for \code{MRGVARDataR()}}
\usage{
MRGVARDataR(...)
}
\description{
This is kept for backwards compatibility. Please use \code{mrgvar_data_r()} instead.
}
