% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{MRGVARest}
\alias{MRGVARest}
\title{Legacy alias for \code{MRGVARest()}}
\usage{
MRGVARest(...)
}
\description{
This is kept for backwards compatibility. Please use \code{mrgvar_estimate()} instead.
}
