% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{mrvar_estimate}
\alias{mrvar_estimate}
\title{Estimate a multi-regime VAR model}
\usage{
mrvar_estimate(res)
}
\arguments{
\item{res}{MRVAR-style data object, typically the output of \code{mrvar_data()},
containing at least \code{n}, \code{S}, \code{p}, \code{type}, \code{Y}, \code{SESVI}, \code{TH}, and \code{d},
and optionally \code{X}.}
}
\value{
The input MRVAR object augmented with estimated parameters and test
statistics.
}
\description{
Estimates the parameters of an MRVAR system from supplied data and returns
the fitted regime-specific VAR objects together with summary statistics.
}
\examples{
p = matrix(c(3,3,0,0),2,2)
res_d = mrvar_data(n=2,p=p,T=200,S=2,SESVI=1,type="const")
colnames(res_d$Y) = c("Y","P")
spectral_radius(res_d$B[,,,1])
spectral_radius(res_d$B[,,,2])
res_e = mrvar_estimate(res=res_d)
res_e$Summary

}
