% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{MRVAR_Select}
\alias{MRVAR_Select}
\title{Legacy alias for \code{MRVAR_Select()}}
\usage{
MRVAR_Select(...)
}
\description{
This is kept for backwards compatibility. Please use \code{mrvar_select_()} instead.
}
