% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{mrvar_select__summary}
\alias{mrvar_select__summary}
\title{Summarize MRVAR model selection results}
\usage{
mrvar_select__summary(Sel)
}
\arguments{
\item{Sel}{Numeric matrix returned by \code{mrvar_select_()}.}
}
\value{
A list of length 2. The first element contains the specification
minimizing BIC, and the second contains the specification minimizing AIC.
}
\description{
Extract the specifications that minimize BIC and AIC from the output of
\code{mrvar_select_()}.
}
\examples{
res_d = mrvar_data(n=4,p=matrix(c(2,1,2,2,0,0,0,0),4,2),T=800,S=2,SESVI=1)
res_e = mrvar_estimate(res=res_d)
TH_v = c(0,0.0)
L_v = c(5,5)
Sel = mrvar_select_(res=res_e,L_V=L_v,TH_V=TH_v)
summary_result = mrvar_select__summary(Sel)
summary_result[[1]]  # Optimal model by BIC
summary_result[[2]]  # Optimal model by AIC

}
