% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{MRVAR_Select_Summary}
\alias{MRVAR_Select_Summary}
\title{Legacy alias for \code{MRVAR_Select_Summary()}}
\usage{
MRVAR_Select_Summary(...)
}
\description{
This is kept for backwards compatibility. Please use \code{mrvar_select__summary()} instead.
}
