% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{MRVARData}
\alias{MRVARData}
\title{Legacy alias for \code{MRVARData()}}
\usage{
MRVARData(...)
}
\description{
This is kept for backwards compatibility. Please use \code{mrvar_data()} instead.
}
