% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{MRVARest}
\alias{MRVARest}
\title{Legacy alias for \code{MRVARest()}}
\usage{
MRVARest(...)
}
\description{
This is kept for backwards compatibility. Please use \code{mrvar_estimate()} instead.
}
