% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{NoutofT}
\alias{NoutofT}
\title{Legacy alias for \code{NoutofT()}}
\usage{
NoutofT(...)
}
\description{
This is kept for backwards compatibility. Please use \code{sample_n_out_of_t()} instead.
}
