% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rnorm_sigma}
\alias{rnorm_sigma}
\title{This function will generate iid multivariate normal random time series.}
\usage{
rnorm_sigma(T, sigma)
}
\arguments{
\item{T}{Length of the generated time series}

\item{sigma}{An (n x n) covariance matrix of the normal series}
}
\value{
T x n matrix of iid normal time series
}
\description{
This function will generate iid multivariate normal random time series.
}
\keyword{internal}
