% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{rnormSIGMA}
\alias{rnormSIGMA}
\title{Legacy alias for \code{rnormSIGMA()}}
\usage{
rnormSIGMA(...)
}
\description{
This is kept for backwards compatibility. Please use \code{rnorm_sigma()} instead.
}
