% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{rnormSIGMA_cond}
\alias{rnormSIGMA_cond}
\title{Legacy alias for \code{rnormSIGMA_cond()}}
\usage{
rnormSIGMA_cond(...)
}
\description{
This is kept for backwards compatibility. Please use \code{rnorm_sigma_cond()} instead.
}
