% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{Roots2coef}
\alias{Roots2coef}
\title{Legacy alias for \code{Roots2coef()}}
\usage{
Roots2coef(...)
}
\description{
This is kept for backwards compatibility. Please use \code{roots_to_coef()} instead.
}
