% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roots_to_coef}
\alias{roots_to_coef}
\title{Given the roots of a characteristic polynom in lags, output the coefficients of the corresponding AR process.}
\usage{
roots_to_coef(p, r_p)
}
\arguments{
\item{p}{The lag length}

\item{r_p}{A p-vector of roots outside the unit circle}
}
\value{
A vector of AR coefficients
}
\description{
Given the roots of a characteristic polynom in lags, output the coefficients of the corresponding AR process.
}
\keyword{internal}
\keyword{roots_to_coef(3,c(1.1,1.2,1.3))}
