% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{RZSt2VECM}
\alias{RZSt2VECM}
\title{Legacy alias for \code{RZSt2VECM()}}
\usage{
RZSt2VECM(...)
}
\description{
This is kept for backwards compatibility. Please use \code{rz_st2_vecm()} instead.
}
