% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shift_z2}
\alias{shift_z2}
\title{Shift common-factor regressors to the end of Z2}
\usage{
shift_z2(Z2, kz, n1, p)
}
\arguments{
\item{Z2}{Regressor matrix to be shifted.}

\item{kz}{Number of common exogenous factors.}

\item{n1}{Number of domestic variables.}

\item{p}{Lag length of domestic regressors in \code{Z2} (also the lag length of common factors).}
}
\value{
A regressor matrix with common-factor columns shifted to the end (when applicable).
}
\description{
Reorders the auxiliary-regression regressor matrix \code{Z2} by moving the lagged
exogenous common-factor blocks (if present) to the end of the matrix. If there are
no common factors (\code{kz = 0}) or no lags (\code{p = 0}), \code{Z2} is returned unchanged.
}
\details{
This is used when the regressor matrix is stacked by lag as
(domestic variables, common factors) for each lag, and a downstream routine expects
domestic regressors first, followed by all common-factor regressors.
}
\examples{
set.seed(1)
T <- 10
n1 <- 2
kz <- 1
p <- 3

# Z2 columns are arranged by lag blocks: (n1 domestic + kz factors) per lag
Z2 <- matrix(rnorm(T * (n1 + kz) * p), nrow = T)
Z2s <- shift_z2(Z2 = Z2, kz = kz, n1 = n1, p = p)
dim(Z2s)

}
\keyword{internal}
