% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{SigmaNPDSelectR}
\alias{SigmaNPDSelectR}
\title{Legacy alias for \code{SigmaNPDSelectR()}}
\usage{
SigmaNPDSelectR(...)
}
\description{
This is kept for backwards compatibility. Please use \code{sigma_npd_select_r()} instead.
}
