% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{summaryCIVAR}
\alias{summaryCIVAR}
\title{Legacy alias for \code{summaryCIVAR()}}
\usage{
summaryCIVAR(...)
}
\description{
This is kept for backwards compatibility. Please use \code{summary_civar()} instead.
}
