% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{VARB_commtrend}
\alias{VARB_commtrend}
\title{Legacy alias for \code{VARB_commtrend()}}
\usage{
VARB_commtrend(...)
}
\description{
This is kept for backwards compatibility. Please use \code{varb_commtrend()} instead.
}
