% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{VECM2VARm}
\alias{VECM2VARm}
\title{Legacy alias for \code{VECM2VARm()}}
\usage{
VECM2VARm(...)
}
\description{
This is kept for backwards compatibility. Please use \code{vecm2_va_rm()} instead.
}
