% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_d.r
\name{AB_CIVARTest}
\alias{AB_CIVARTest}
\title{Test on restrictions in the cointegration space}
\usage{
AB_CIVARTest(R0, R1, G, H, h, alpha, beta, Omega, Y0, Z1, Z2)
}
\arguments{
\item{R0}{: T x n matrix containing the residuals of the auxiliary regression of dy_t on Z2_t}

\item{R1}{: T x n matrix containing the residuals of the auxiliary regression of y_(t-1) on Z2_t}

\item{G}{: The restriction matrix on alpha}

\item{H}{: The restriction matrix on beta}

\item{h}{: The restriction vector on beta ( free-varying parameters of beta )}

\item{alpha}{: n x crk matrix containing initial values of the adjustment parameters}

\item{beta}{: n x crk matrix containing initial values of the cointegration vectors}

\item{Omega}{: n x n matrix containing initials of the covariance matrix}

\item{Y0}{: T x n matrix of dY_t}

\item{Z1}{: T x n matrix containing the lagged cointegrated VAR process.}

\item{Z2}{: T x (n(p-1)+#det) matrix containing the lagged stationary regressors of the auxiliary regression}
}
\value{
a list containing (VECMRS, beter, alphar, LSKOEFR, LR, and error)
}
\description{
This function estimates constrained VECM using the iteration procedure proposed in Boswijk and Doornik (2003)
}
\section{Details}{

This function runs a likelihood ratio test of linear restrictions on \eqn{\alpha} and \eqn{\beta} in a CIVAR model in the following form:
\deqn{vec(\alpha') = G \psi,  vec(\beta) = H\phi + h}

example 1 (restrictions on alpha) test of exogeneity: one weakly exogenous variable
vec(alpha) is n\emph{crk x 1  vector
G          is n}crk x (n-k)\emph{crk matrix  (k is the number of weakly exogenous)
psi        (n-k)\emph{crk x 1 vector
vec(beta)  is n}crk x 1 vector
H          is n}crk x n\emph{crk identity matrix
phi        n}crk x 1 vector
h          n*crk x 1 zero matrix implying ver(beta) = phi.
(H is identity and h is zero vector implies only restrictions on alpha)

example 2 (restrictions on beta ) test of PPP
vec(alpha) is n\emph{crk x 1 vector
G          is n}crk x n\emph{crk identity matrix, implying no restriction on alpha
psi        n}crk x 1 vector
vec(beta)  n\emph{crk x 1 vector
H          n}crk x 2 matrix that picks out the elements under restrictions
two columns out of the identity matrix.
ones in a row of H and zero in the corresponding h implies non-restricted beta.
phi        2 x 1  vector
h          n*crk x 1  non zero elements in this vector together with the zero elements in
the corresponding row in H are the normalization conditions.
(H is identity and h is zero vector implies only restrictions on alpha)
}

\keyword{internal}
