% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alpha_perpf}
\alias{alpha_perpf}
\title{Compute an orthogonal complement of a matrix}
\usage{
alpha_perpf(alpha = (alpha))
}
\arguments{
\item{alpha}{Matrix with full-column rank.}
}
\value{
A matrix whose columns form a basis for the orthogonal complement of
\code{alpha}.
}
\description{
Construct a matrix whose columns span the orthogonal complement of the
column space of \code{alpha}.
}
\examples{
set.seed(1)
n <- 5
r <- 2
alpha <- matrix(rnorm(n * r), nrow = n, ncol = r)
alpha_perp <- alpha_perpf(alpha)

# Check orthogonality (should be close to zero)
t(alpha) \%*\% alpha_perp

}
\keyword{internal}
