% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bo2_ao}
\alias{bo2_ao}
\title{This function calculates Ao coefficients from Bo with common trends.}
\usage{
bo2_ao(m, p, T, r, Ncommtrend, n, S, r_npo, DFYflag = NA)
}
\arguments{
\item{m}{Number of variables}

\item{p}{Lags}

\item{T}{Number of observations}

\item{r}{Number of stochastic trends}

\item{Ncommtrend}{Number of the common trends}

\item{n}{Number of countries}

\item{S}{Number of regimes}

\item{r_npo}{The matrix containing the roots of the characteristic polynomial}

\item{DFYflag}{switching indicator whether the foreign variables enter the cointegration space}
}
\value{
Parameter matrix of the foreign variables in the country equation of the GVAR model
}
\description{
This function calculates Ao coefficients from Bo with common trends.
}
\keyword{internal}
