% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cia2_a}
\alias{cia2_a}
\title{Convert foreign-block VECM coefficients to VAR levels form}
\usage{
cia2_a(CIA)
}
\arguments{
\item{CIA}{Three-dimensional array of VECM coefficients.}
}
\value{
A three-dimensional array of VAR coefficients.
}
\description{
Transform an array of foreign-block VECM coefficients into the corresponding
VAR coefficient arrays in levels.
}
\examples{
# Example with m = 2 variables and L = 3 lags in the array
set.seed(1)
m <- 2
L <- 3
CIA <- array(rnorm(m * m * L), dim = c(m, m, L))
A <- cia2_a(CIA)
dim(A)
A[,,1]

}
\keyword{internal}
