% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cib3_b}
\alias{cib3_b}
\title{Convert VECM coefficient arrays to VAR coefficient arrays}
\usage{
cib3_b(CIB)
}
\arguments{
\item{CIB}{An (m, m, L) array of coefficient matrices of a VECM with lag L-1.}
}
\value{
B An (m, m, L) array of coefficient matrices of a VAR with lag L.
}
\description{
Transforms an array of VECM coefficient matrices (with lag order L-1) into the
corresponding VAR coefficient matrices (with lag order L). The input and output
are both 3D arrays, where the third dimension indexes lags.
}
\examples{
# Example with m = 3 variables and L = 2 (so VECM lag is L-1 = 1)
set.seed(1)
m <- 3
L <- 2
CIB <- array(rnorm(m * m * L), dim = c(m, m, L))
B <- cib3_b(CIB)
dim(B)
B[,,1]

}
\keyword{internal}
