% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{cigvar_select_}
\alias{cigvar_select_}
\title{CIGVAR Model Selection Function}
\usage{
cigvar_select_(res, L_V = L_V)
}
\arguments{
\item{res}{: an object generated from CIGVARest.}

\item{L_V}{: a 2 components vector containing the maxima of the domestic lag and the foreign lag, respectively.}
}
\value{
a matrix with different lag specifications and the corresponding information criteria.
}
\description{
This function performs model selection for the CIGVAR (Cointegrated Global Vector Autoregression) framework by calculating AIC and BIC criteria for various lag specifications of domestic and foreign variables for multiple countries.
}
\examples{

n = 5
p = (1:15)*0; dim(p) = c(5,3)
p[,1] = 3; p[,2]=3;
p[2,1] = 2;  p[3,1] = 2
res_d = cigvar_data(m=2,n=5,p=p,T=5000,type="const",DFYflag=0)
res_d$r_npo
res_d$check

plot(ts(res_d$Y[1:100,1:10]))

res_e = cigvar_estimate(res_d)
res_e$Summary
res_e$crk = c(1,1,1,1,1)
L_V = c(4,4)
CIGVARSelecte = cigvar_select_(res=res_e,L_V=c(4,4))
CIGVARSelecte[which.min(CIGVARSelecte[,17]),]
CIGVARSelecte[which.min(CIGVARSelecte[,19]),]

}
