% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{girf_MRGVAR_RM}
\alias{girf_MRGVAR_RM}
\title{Legacy alias for \code{girf_MRGVAR_RM()}}
\usage{
girf_MRGVAR_RM(...)
}
\description{
This is kept for backwards compatibility. Please use \code{girf_mrgvar_rm()} instead.
}
