% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{girf_MRGVAR_RM_CB}
\alias{girf_MRGVAR_RM_CB}
\title{Legacy alias for \code{girf_MRGVAR_RM_CB()}}
\usage{
girf_MRGVAR_RM_CB(...)
}
\description{
This is kept for backwards compatibility. Please use \code{girf_mrgvar_rm_cb()} instead.
}
