% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{girf_mrvar_rm_cb}
\alias{girf_mrvar_rm_cb}
\title{Calculate Generalized Impulse Response Functions with Bootstrapped Confidence Intervals for MRVAR Models}
\usage{
girf_mrvar_rm_cb(
  RES,
  shock,
  R,
  nstep,
  Omega_hist = NA,
  resid_method = "parametric",
  conf_level,
  N
)
}
\arguments{
\item{RES}{: an estimated MRVAR object.}

\item{shock}{: an n-vector containing the shocks as impulse.}

\item{R}{: number of runs to integrate out the random effects in order to obtain the means (see equation above).}

\item{nstep}{: length of the impulse response function.}

\item{Omega_hist}{: a (P x n) matrix of initial values, from which the impulse response functions start.
Omega_hist determines from which regime the impulse response functions start.}

\item{resid_method}{: resid_method = c("resid", "parametric"), it generates the random residuals from
residuals bootstrap or parametric bootstrap.}

\item{conf_level}{: a vector containing the levels of confidence.}

\item{N}{: the number of bootstrap runs in containing the bootstrapped confidence intervals.}
}
\value{
an (n, n, nstep+1, 3) array containing the impulse response functions with lower and upper confidence bands.
The rows represent responses, and the columns represent impulses.
}
\description{
This function calculates the generalized impulse response functions (GIRF) of an estimated MRVAR(n,p,S)
for a given shock vector. It also provides bootstrapped confidence intervals for the impulse response functions.
}
\details{
\subsection{GIRF(shock=SHCK) = mean(Y(resid)) - mean(Y(SHCK))}{

See H.H. Pesaran and Y. Shin (1998) Generalized impulse response analysis in linear multivariate models,
Economics Letters, 58(1) p. 17-29, and G. Koop, M. H. Pesaran, and S. M. Potter (1996),
Impulse response analysis in nonlinear multivariate models, Journal of Econometrics, 74 (1996) 119-74.
}
}
\examples{
p = matrix(c(2,1,0,0),2,2)
res_d = mrvar_data(n=2,p=p,T=300,S=2,SESVI=1,type="none")
max(res_d$Y)
colnames(res_d$Y) = c("R","P")
res_e = mrvar_estimate(res=res_d)
RF3 = girf_mrvar_rm(RES=res_e,shock=c(1,1),R = 200,nstep=20,Omega_hist=NA,resid_method="parametric")
RF4 = girf_mrvar_rm_cb(RES=res_e, shock=c(1,1), R=200, nstep=20, Omega_hist=NA,
                        resid_method="parametric", conf_level=c(0.05,0.95), N=100)
IRF_list <- plot_irf(RF4)
}
