% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{infer_deterministic_type}
\alias{infer_deterministic_type}
\title{This function will output type according to specification of Co and EXOG and type.}
\usage{
infer_deterministic_type(Co, EXOG, type)
}
\arguments{
\item{Co}{The coefficients of the deterministic components}

\item{EXOG}{The exogenous variables}

\item{type}{The type of the deterministic specification for comparison}
}
\value{
type
}
\description{
This function will output type according to specification of Co and EXOG and type.
}
\examples{
infer_deterministic_type(Co=matrix(c(0,0,1,1),2,2),EXOG= c(1:10),type="exog0")
infer_deterministic_type(Co=matrix(c(1,1,1,1),2,2),EXOG= c(1:10),type="exog1")
infer_deterministic_type(Co=matrix(c(1,1,1,1,0,0),2,3),EXOG= c(1:10),type=NA)
infer_deterministic_type(Co=c(1,1),EXOG= NA,type=NA)
infer_deterministic_type(Co=c(0,0),EXOG= NA,type=NA)
var_data(n=2,p=2,T=100,Co=matrix(c(2,2,1,1),2,2),type="exog0",X=c(1:100))
var_data(n=2,p=2,T=100,Co=matrix(c(1,1,1,1),2,2),type="exog1",X=c(1:100))
}
