% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf_gvar}
\alias{irf_gvar}
\title{This function generates impulse response functions of an estimated GVAR}
\usage{
irf_gvar(res, nstep, comb, irf = c("gen", "chol", "chol1", "gen1", "comb1"))
}
\arguments{
\item{res}{: an output of gvar_estimate}

\item{nstep}{: length of the impulse response functions}

\item{comb}{: an mn vector specifying combined impulse such as global shocks, regional shocks, or concerted actions.}

\item{irf}{: types of the impulse response functions. irf=c("gen","chol","chol1","gen1","comb1"), gen for generalized IRF with one standard deviation shocks, gen1 for generalized IRF with one unit impulse, chol for IRF with Cholezky decomposition of the covariance matrix, chol1 for Cholezky decomposition with one unit impulse, comb1 for concerted action with one unit impulse.}
}
\value{
an (mn,mn,nstep) array containing the IRF with columns representing the impulses rows representing the responses.
}
\description{
This function generates impulse response functions of an estimated GVAR
}
\examples{


X1 = matrix(1,200,1)
X2 = matrix(rnorm(200),200,1)
X3 = matrix(rnorm(200),200,1)
X4 = matrix(rnorm(200),200,1)
X  = cbind(X1,X2,X3,X4)
dim(X) = c(200,1,4)

n = 4
p = (1:12)*0; dim(p) = c(4,3);p[,1] = 2; p[,2]=2; p[,3]=1;
res_d = gvar_data(m=2,n=4,p=p,T=200,type="exog0",X=X)
res_e = gvar_estimate(res=res_d)
IRF = irf_gvar(res_e,nstep=10,comb=NA,irf="gen")
IRF_CB = irf_gvar_cb(res_e,nstep=10,comb=NA,irf="gen",runs=200,conf=c(0.05,0.95))
dim(IRF_CB)
IRF_g = plot_irf(IRF_CB,Names=NA,response=c(1,4),impulse=c(1,2,3,4), ncol=4)

}
\keyword{internal}
