% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{irf_MRCIGVAR}
\alias{irf_MRCIGVAR}
\title{Legacy alias for \code{irf_MRCIGVAR()}}
\usage{
irf_MRCIGVAR(...)
}
\description{
This is kept for backwards compatibility. Please use \code{irf_mrcigvar()} instead.
}
