% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{irf_MRCIVAR_CB}
\alias{irf_MRCIVAR_CB}
\title{Legacy alias for \code{irf_MRCIVAR_CB()}}
\usage{
irf_MRCIVAR_CB(...)
}
\description{
This is kept for backwards compatibility. Please use \code{irf_mrcivar_cb()} instead.
}
