% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{irf_MRGVAR}
\alias{irf_MRGVAR}
\title{Legacy alias for \code{irf_MRGVAR()}}
\usage{
irf_MRGVAR(...)
}
\description{
This is kept for backwards compatibility. Please use \code{irf_mrgvar()} instead.
}
