% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf_mrgvar}
\alias{irf_mrgvar}
\title{Generate Impulse Response Functions for MRGVAR Models}
\usage{
irf_mrgvar(
  res = res,
  state = state,
  nstep = nstep,
  comb = comb,
  irf = c("gen", "chol", "chol1", "gen1", "genN1", "comb1", "smat"),
  G = NA,
  smat = NA,
  sigmaNPDS = NA
)
}
\arguments{
\item{res}{a list of estimated MRGVAR as output of mrgvar_estimate}

\item{state}{an n vector specifying the specific state for each country}

\item{nstep}{the length of impulse response function}

\item{comb}{a vector specify the concerted action in policy-simulation impulse response function}

\item{irf}{types of the impulse response irf=c("gen","chol","chol1","gen1","genN1","comb1","smat")}

\item{G}{For permanent and transitory decomposition}

\item{smat}{For explicit structural decomposition of the correlated shocks}

\item{sigmaNPDS}{the state-dependent covariance matrix}
}
\value{
an (mn,mn,nstep) array containing the IRF with columns representing the impulses and rows the responses
}
\description{
This function computes the impulse response functions (IRFs) of an estimated MRGVAR model,
allowing for various types of IRFs and handling different configurations of shocks and responses
across multiple regimes and countries.
}
\examples{
## case of n = 2, m = 2, S = 2     ## m: number of variables, n: number of countries
p = rep(1,12); dim(p) = c(2,3,2)
p[1,1,2] = 2; p[2,2,2]=2; p[,3,] = 0
TH = c(1:2)*0; dim(TH) = c(1,2)
res_d <- mrgvar_data(m=2,n=2,p=p,TH=TH,T=100,S=2,SESVI=c(1,3),type="const")
max(res_d$Y)

### estimation of the MRGVAR model
colnames(res_d$Y) = c("P","Q","Pa","Qa")
res_e = mrgvar_estimate(res=res_d)
res_e$Summary

IRF  = irf_mrgvar(res=res_e,nstep=10,comb=NA,state=c(1,1),irf="gen1")

}
\keyword{internal}
