% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{irf_MRGVAR_CB}
\alias{irf_MRGVAR_CB}
\title{Legacy alias for \code{irf_MRGVAR_CB()}}
\usage{
irf_MRGVAR_CB(...)
}
\description{
This is kept for backwards compatibility. Please use \code{irf_mrgvar_cb()} instead.
}
