% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{irf_MRVAR_CB}
\alias{irf_MRVAR_CB}
\title{Legacy alias for \code{irf_MRVAR_CB()}}
\usage{
irf_MRVAR_CB(...)
}
\description{
This is kept for backwards compatibility. Please use \code{irf_mrvar_cb()} instead.
}
