% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{irf_MRVAR_NM}
\alias{irf_MRVAR_NM}
\title{Legacy alias for \code{irf_MRVAR_NM()}}
\usage{
irf_MRVAR_NM(...)
}
\description{
This is kept for backwards compatibility. Please use \code{irf_mrvar_nm()} instead.
}
