% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{irf_VAR_CB}
\alias{irf_VAR_CB}
\title{Legacy alias for \code{irf_VAR_CB()}}
\usage{
irf_VAR_CB(...)
}
\description{
This is kept for backwards compatibility. Please use \code{irf_var_cb()} instead.
}
