% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{mrcigvar_data_r}
\alias{mrcigvar_data_r}
\title{Regenerate data from an estimated MRCIGVAR object}
\usage{
mrcigvar_data_r(res)
}
\arguments{
\item{res}{: an output of mrcigvar_estimate}
}
\value{
: an MRCIGVAR object.
}
\description{
Regenerates synthetic data from an estimated \code{mrcigvar_estimate()} result,
typically to obtain a sample large enough for downstream estimation or
bootstrap-style procedures.
}
\keyword{internal}
