% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{mrcigvar_data_r}
\alias{mrcigvar_data_r}
\title{This function will generate data from a MRCIGVAR object. It will generate enough data for estimation purpose.}
\usage{
mrcigvar_data_r(res)
}
\arguments{
\item{res}{: an output of mrcigvar_estimate}
}
\value{
: an MRCIGVAR object.
}
\description{
This function will generate data from a MRCIGVAR object. It will generate enough data for estimation purpose.
}
\keyword{internal}
