% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{mrcigvar_select_}
\alias{mrcigvar_select_}
\title{Select lag and threshold specifications for an MRCIGVAR model}
\usage{
mrcigvar_select_(res, L_V = L_V, TH_V = TH_V)
}
\arguments{
\item{res}{An MRCIGVAR object, typically returned by \code{mrcigvar_data()} or
\code{mrcigvar_estimate()}.}

\item{L_V}{Numeric vector of length 2 giving the maximum domestic and
foreign lag orders considered in the search.}

\item{TH_V}{Numeric vector of candidate threshold values.}
}
\value{
A numeric matrix containing the evaluated lag and threshold
specifications together with the information criteria returned by the
estimator.
}
\description{
Evaluate alternative domestic lag, foreign lag, and threshold
specifications for an MRCIGVAR model and return the resulting information
criteria.
}
\examples{
m = 2                  ## m: number of variables, n: number of countries
n = 4
p = c(2,2,2,2,2,2,2,2,2,2,0,0,0,0,0,2,2,2,2,2,2,2,2,2,2,0,0,0,0,0); dim(p) = c(5,3,2)
p = p[1:4,,]; p[,1:2,2] = 3

TH = c(1:4)*0; dim(TH) = c(1,4)

res_d <- mrcigvar_data(m=2,n=4,p=p,TH=TH,T=400,S=2, SESVI=c(1,3,5,7),r=rep(1,4))

res_em = mrcigvar_estimate(res=res_d)

L_v = c(3,3)
TH_v = c(0,0.1)
res_d$p

MRCIGVARSelect = mrcigvar_select_(res=res_d,L_V=L_v,TH_V=TH_v)
dim(MRCIGVARSelect)
MRCIGVARSelect[which.min(MRCIGVARSelect[,28]),]

}
