% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{mrcivar_select_m}
\alias{mrcivar_select_m}
\title{MRCIVAR Model Selection Function}
\usage{
mrcivar_select_m(res = res, L_V = L_V, TH_V = TH_V)
}
\description{
This function performs model selection for the MRCIVAR framework by calculating AIC and BIC criteria for various lag specifications and threshold values.
}
\examples{

Sigma = 1:(4*4*2)
dim(Sigma) = c(4,4,2)
Sigma[,,1] = diag(4)
Sigma[,,2] = diag(4)
p=matrix(0,2,2)
p[,1] = c(3,2)

res_d = mrcivar_data_m(n=4,p=p,T=2610,S=2,SESVI=1,TH=0,Sigmao=Sigma,type="const",r=1)
res_e = mrcivar_estimatem1(res=res_d)

TH_v = c(0,0.1)
L_v = c(6,6)

Selm = mrcivar_select_m(res=res_e,L_V=L_v,TH_V=TH_v)
mrvar_select__summary(Selm)

}
