% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{mrgvar_data_r}
\alias{mrgvar_data_r}
\title{This function will generate data from a MRGVAR object. It will generate enough data for estimation purpose.}
\usage{
mrgvar_data_r(res)
}
\arguments{
\item{res}{: an output of mrgvar_estimate}
}
\value{
an MRGVAR object.
}
\description{
This function will generate data from a MRGVAR object. It will generate enough data for estimation purpose.
}
\keyword{internal}
