% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{mrvar_estimate}
\alias{mrvar_estimate}
\title{This function estimates the parameters of a specified MRVAR(n,p,S) based on provided data.}
\usage{
mrvar_estimate(res)
}
\arguments{
\item{res}{:an object of MRVAR that is an output of mrvar_data including at least: n, S, p, type, Y, SESVI, TH, d, and optionally X.}
}
\value{
an MRVAR object that is a list containing estimated parameters and some test statistics.
}
\description{
This function estimates the parameters of a specified MRVAR(n,p,S) based on provided data.
}
\examples{
p = matrix(c(3,3,0,0),2,2)
res_d = mrvar_data(n=2,p=p,T=200,S=2,SESVI=1,type="const")
colnames(res_d$Y) = c("Y","P")
spectral_radius(res_d$B[,,,1])
spectral_radius(res_d$B[,,,2])
res_e = mrvar_estimate(res=res_d)
res_e$Summary

}
