% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{mrvar_select_}
\alias{mrvar_select_}
\title{MRVAR Model Selection Function}
\usage{
mrvar_select_(res, L_V = L_V, TH_V = TH_V)
}
\arguments{
\item{res}{: an MRVAR object generated from mrvar_data or estimated from mrvar_estimate.}

\item{L_V}{: a 2 components vector containing the maxima lags in the two regimes, respectively.}

\item{TH_V}{: a vector containing the possible threshold values over which the model selection criteria values will be calculated.}
}
\value{
a matrix with different lag specifications and threshold values as well as the corresponding information criterion values.
}
\description{
This function performs model selection for the MRVAR (Markov Regime-Switching VAR) framework by calculating AIC, BIC, and other information criteria for various lag specifications and threshold values across regimes.
}
\examples{
res_d = mrvar_data(n=4,p=matrix(c(2,1,2,2,0,0,0,0),4,2),T=800,S=2,SESVI=1)
max(res_d$Y)
colnames(res_d$Y) = c("P","Y","R","U")
res_e = mrvar_estimate(res=res_d)
TH_v = c(0,0.0)
L_v = c(5,5)
Sel = mrvar_select_(res=res_e,L_V=L_v,TH_V=TH_v)
mrvar_select__summary(Sel)

}
