% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{plot_irf}
\alias{plot_irf}
\title{Plot Impulse Response Functions with Confidence Bands}
\usage{
plot_irf(
  IRF_CB = IRF_CB,
  Names = NA,
  INames = NA,
  response = c(1:n),
  impulse = c(1:n),
  ncol = n
)
}
\arguments{
\item{IRF_CB}{An (n x n x L x 3) array of impulse response function with confidence bands}

\item{Names}{An n-vector of strings of the variable names}

\item{INames}{An n-vector of string of the impulse names}

\item{response}{A vector of impulse indices}

\item{impulse}{A vector of response indices}

\item{ncol}{Number of columns of impulse response functions in the plot}
}
\value{
An ggplot object of impulse response functions
}
\description{
This function generates a list of ggplot objects for visualizing impulse response functions
along with their confidence bands. It allows for customization of variable names and impulse
indices for better clarity in the plots.
}
\keyword{internal}
