% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shift_z2m}
\alias{shift_z2m}
\title{Shift common-factor regressors to the end of a mixed lag matrix}
\usage{
shift_z2m(Z2, kz, n1, p, p2)
}
\arguments{
\item{Z2}{Regressor matrix to reorder.}

\item{kz}{Integer scalar giving the number of common exogenous factors.}

\item{n1}{Integer scalar giving the number of domestic variables.}

\item{p}{Integer scalar giving the domestic lag length.}

\item{p2}{Integer scalar giving the foreign lag length.}
}
\value{
A reordered regressor matrix.
}
\description{
Reorder a regressor matrix so that lagged common-factor columns are moved to
the end when both domestic and foreign lag blocks are present.
}
\examples{
set.seed(1)
T <- 10
n1 <- 2
kz <- 1
p <- 2
p2 <- 2

# Fabricate a Z2 with enough columns for the intended indexing pattern
# (Exact layout depends on how Z2 is constructed upstream.)
ncol_Z2 <- 2 * (p * (n1 + kz) + p2 * n1)
Z2 <- matrix(rnorm(T * ncol_Z2), nrow = T, ncol = ncol_Z2)

Z2s <- shift_z2m(Z2 = Z2, kz = kz, n1 = n1, p = p, p2 = p2)
dim(Z2s)

}
\keyword{internal}
