% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{sigma_npd}
\alias{sigma_npd}
\title{Compute Near Positive Definite Covariance Matrix for Selected States}
\usage{
sigma_npd(res, StateT)
}
\arguments{
\item{res}{An estimated GVAR object}

\item{StateT}{A vector of selected states of each country}
}
\value{
The covariance matrix of the selected states
}
\description{
Calculates the covariance matrix of residuals for selected states from a GVAR model
and returns its nearest positive definite approximation.
}
\examples{
\dontrun{
  sigma_npd(gvar_model, c(1, 2, 3))
}

}
\keyword{internal}
