% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary_civar}
\alias{summary_civar}
\title{Summary of Conditional VAR Model}
\usage{
summary_civar(lm1 = lm1, sname = "Z2\\\\[,4:9\\\\]")
}
\arguments{
\item{lm1}{A fitted conditional VAR model object.}

\item{sname}{The elements in column names to be replaced.}
}
\description{
This function provides a summary of a conditional VAR model, allowing for the replacement of specific elements in the column names of the output.
}
\examples{
set.seed(1)
dat <- data.frame(
  y  = rnorm(50),
  x1 = rnorm(50),
  x2 = rnorm(50)
)
model <- lm(y ~ x1 + x2, data = dat)
summary_result <- summary_civar(model, sname ="Z2\\\\[,4:9\\\\]")
print(summary_result)
}
\keyword{internal}
