% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{varb_commtrend}
\alias{varb_commtrend}
\title{This function creates independent common stochastic trends and idiosyncratic trends.}
\usage{
varb_commtrend(m, p, T, r_npo, Ncommtrend, n)
}
\arguments{
\item{m}{number of variables}

\item{p}{lags}

\item{T}{number of observations}

\item{r_npo}{roots of the characteristic polynomial}

\item{Ncommtrend}{number of common trends}

\item{n}{number of countries}
}
\value{
A list containing the parameter matrix of the domestic variables, the cointegration vectors and the adjustment vectors
}
\description{
This function creates independent common stochastic trends and idiosyncratic trends.
}
\keyword{internal}
