% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{varb_commtrend}
\alias{varb_commtrend}
\title{Generate Independent Common Stochastic Trends and Idiosyncratic Trends for VAR Models}
\usage{
varb_commtrend(m, p, T, r_npo, Ncommtrend, n)
}
\arguments{
\item{m}{number of variables}

\item{p}{lags}

\item{T}{number of observations}

\item{r_npo}{roots of the characteristic polynomial}

\item{Ncommtrend}{number of common trends}

\item{n}{number of countries}
}
\value{
A list containing the parameter matrix of the domestic variables,
the cointegration vectors, and the adjustment vectors
}
\description{
This function creates independent common stochastic trends and idiosyncratic trends
for a given number of variables, lags, and observations. It also computes the
parameter matrix of domestic variables, cointegration vectors, and adjustment vectors
based on the specified roots of the characteristic polynomial.
}
\examples{
\dontrun{
m <- 3
p <- matrix(c(1, 2,
              2, 1), ncol = 2, byrow = TRUE)
T <- 100
n <- 2
Pmax <- max(p[, 1])

# Provide valid roots array (see package vignette for recommended settings)
r_npo <- array(runif(m * Pmax * n), dim = c(m, Pmax, n))

Ncommtrend <- 1
result <- varb_commtrend(m, p, T, r_npo, Ncommtrend, n)
print(result)
}

}
\keyword{internal}
