% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-legacy-aliases.R
\name{VARBS_commtrend}
\alias{VARBS_commtrend}
\title{Legacy alias for \code{VARBS_commtrend()}}
\usage{
VARBS_commtrend(...)
}
\description{
This is kept for backwards compatibility. Please use \code{varbs_commtrend()} instead.
}
