% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vecm2_var}
\alias{vecm2_var}
\title{Convert VECM Parameters to VAR Parameters for CIGVAR}
\usage{
vecm2_var(param, beta, q = c(1, 2, 2, 2, 2), s = NA, kz = 0, Dxflag = 0)
}
\arguments{
\item{param}{The parameter of an estimated VECM}

\item{beta}{The cointegration vectors}

\item{q}{A vector specifying different types of VECMs}

\item{s}{The indicator of multi regime VECMs}

\item{kz}{The number of exogenous common factors}

\item{Dxflag}{Indicator whether the foreign variables enter the cointegration space}
}
\value{
A list of (A, B, C) with B the auto regression parameter matrix,
C the parameter of the deterministic components, and A the parameter
matrix of foreign variables for CIGVAR.
}
\description{
This function transforms the estimated parameters of a Vector Error Correction Model (VECM)
into the corresponding parameters for a Conditional VAR (CIGVAR) model. It handles different
types of VECMs and allows for the inclusion of foreign variables and common factors.
}
\examples{
# Example usage of vecm2_var function
set.seed(1)
m <- 3
q <- c(2, 2, 2, 2, 2)
n_rows <- q[1] + sum(q[-1]) * m  # minimum rows needed when kz = 0
param <- matrix(rnorm(n_rows * m), nrow = n_rows, ncol = m)
beta  <- matrix(rnorm(m * q[1]), nrow = m, ncol = q[1])
result <- vecm2_var(param, beta, q = q, s = NA, kz = 0, Dxflag = 0)
str(result)
}
\keyword{internal}
